Factor out frontend/views.py.
