+++ frontend/tests/util.py
@@ -1,0 +1,6 @@
+from frontend.views import FRONTEND_QUERY_DIR, add_graphql_fragments
+
+
+def get_frontend_query(filename) -> str:
+    query = (FRONTEND_QUERY_DIR / filename).read_text()
+    return add_graphql_fragments(query)
