+++ project/tests/test_schema.py
@@ -6,7 +6,7 @@ from users.tests.factories import UserFactory
 from project.util import schema_json
 from project.util.testing_util import GraphQLTestingPal
 from project.schema_base import update_last_queried_phone_number, PhoneNumberAccountStatus
-from .util import get_frontend_query
+from frontend.tests.util import get_frontend_query
 
 EXAMPLE_DEPRECATED_FIELD = 'exampleDeprecatedField'
