+++ rh/tests/test_schema.py
@@ -1,4 +1,4 @@
-from project.tests.util import get_frontend_query
+from frontend.tests.util import get_frontend_query
 from rh.tests.factories import RentalHistoryRequestFactory
 from rh.schema import get_slack_notify_text
 from rh.models import RentalHistoryRequest
